Add Main class
